```julia
using StochasticDiffEq, SimpleDiffEq, BridgeDiffEq
using BenchmarkTools

α=1
β=1
u0=1/2
f(u,p,t) = u
g(u,p,t) = u
dt = 1//2^(4)
tspan = (0.0,1.0)
prob = SDEProblem(f,g,u0,(0.0,1.0))
sol = solve(prob,EM(),dt=dt)

@btime sol = solve(prob,EM(),dt=dt)
@btime sol = solve(prob,BridgeEuler(),dt=dt)
@btime sol = solve(prob,SimpleEM(),dt=dt)

using Profile
solve(prob,EM(),dt=dt)
@profile for i in 1:200000 solve(prob,EM(),dt=dt) end
Juno.profiler()
Profile.clear()

Before: 9.799 μs (90 allocations: 20.72 KiB)
After: 1.810 μs (35 allocations: 2.41 KiB
No error check: 1.400 μs (32 allocations: 2.31 KiB)
Bridge: 864.151 ns (13 allocations: 1.38 KiB)
SOL: 505.155 ns (6 allocations: 656 bytes)
```

Note that by the time you get to Lorenz this difference has eroded:

```julia
using StaticArrays

function lorenz(u,p,t)
 SA[10.0(u[2]-u[1]),u[1]*(28.0-u[3]) - u[2],u[1]*u[2] - (8/3)*u[3]]
end

function σ_lorenz(u,p,t)
 SA[3.0,3.0,3.0]
end

prob_sde_lorenz = SDEProblem{false}(lorenz,σ_lorenz,SA[1.0,0.0,0.0],(0.0,1.0))
@btime sol = solve(prob_sde_lorenz,EM(),dt=0.05)
@btime sol = solve(prob_sde_lorenz,BridgeEuler(),dt=0.05)

EM: 1.950 μs (34 allocations: 2.92 KiB)
BridgeEuler: 1.967 μs (14 allocations: 2.25 KiB)
```

so it's for extremely small equations.

The remaining portions were analyzed. `No error check` is with `check_error!(integrator)` commented out. It turns out that one of the most expensive operations at this size is `any(isnan,u)`. Then I prepared a branch https://github.com/SciML/StochasticDiffEq.jl/pull/303 which gets rid of the tstops, saveat, and d_discontinuities mechanisms, that takes it down another notch. The remaining portions are just the small allocations due to the output array pieces, which are essentially size zero so they have an O(1) cost here, but are hard to get rid of since constant prop isn't always able to infer if it's `()`, so it's safest just to leave it. 

For the most part this shouldn't effect much normal usage given the Lorenz result, but it should help compile times and test times.